code/pyproject.toml
[tool.poetry]
name = "sample"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
kubernetes = "^25.3.0"
pyyaml = "^6.0"
kopf = "^1.36.0"


[tool.poetry.group.dev.dependencies]
mypy = "^0.991"
autopep8 = "^2.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
